fix: don't delete from s3 if --all-but-latest is provided (#398)
* fix: don't delete from s3 if --all-but-latest is provided

* changeset

+++ packages/provider-s3/src/lib/providerS3.ts
@@ -108,9 +108,15 @@ export class S3BuildCache implements RemoteBuildCache {
 
   async delete({
     artifactName,
+    skipLatest,
   }: {
     artifactName: string;
+    skipLatest?: boolean;
   }): Promise<RemoteArtifact[]> {
+    if (skipLatest) {
+      // Artifacts on S3 are unique by name, so skipping latest means we don't delete anything
+      return [];
+    }
     await this.s3.send(
       new clientS3.DeleteObjectCommand({
         Bucket: this.bucket,
